% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAssessment-package.R
\docType{package}
\name{DataAssessment-package}
\alias{DataAssessment}
\alias{DataAssessment-package}
\title{DataAssessment: OHDSI data assessment of a prostate cancer treatment cohort cascade}
\description{
See \code{\link{execute}} for the main entry point and \code{extras/codeToRun.R}
for a script to run the study at a site.
}
\author{
\strong{Maintainer}: Andrei Kiselev \email{kiselev.eco@gmail.com}

Authors:
\itemize{
  \item Andrei Kiselev \email{kiselev.eco@gmail.com}
}

}
\keyword{internal}
